Favor sdkv2 (#57)
* Favor SDK v2

This PR does two things. The first is that in AwsShim, it changes the behavior to favor AWS SDK v2 if it's in the load path over v1. Presumably if you have both in your load path you want the most recent. The other thing that this PR does is allow you to manually specify which version of the SDK to use through the config object. That would look like:

    Humidifier.configure { |config| config.sdk_version = 1 }

* Add docs on specifying the SDK version