Update specs to newest release
This includes fixes for multiple AWS::EC2::SpotFleet resources, as well as changing AWS::IAM::Role AssumeRolePolicyDocument to a Json property. All of these changes came from AWS, so the only change to Humidifier is removing some of the now obsolete fixes to the downloaded JSON in the Fixer class.