Nested conversions and fixing AS/CloudWatch
There are still a couple things wrong with the JSON spec, so this commit fixes a couple of those. This commit also builds in converting "raw" values (i.e., values read from a JSON file or similar source all of the way down the subprop chain. This means that values read in that have nested structures will be properly converted as well.